requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
category-encoders==2.6.3
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.2
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.8
dacite==1.8.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
dateparser==1.1.8
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
Django==4.2.3
dnspython==2.4.2
emoji==2.10.1
executing==1.2.0
faiss-cpu==1.8.0
fastjsonschema==2.16.2
filelock==3.13.1
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.34
google-ai-generativelanguage==0.4.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.61.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2024.2.1
htmlmin==0.1.12
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.4
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
ImageHash==4.3.1
imbalanced-learn==0.12.0
importlib-metadata==6.8.0
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaleido==0.2.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
langsmith==0.1.38
libclang==16.0.6
lightgbm==4.3.0
llama-index-core==0.10.25.post3
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.7
llamaindex-py-client==0.1.15
llvmlite==0.41.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.1
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.5
multimethod==1.11.1
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.14.3
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.14
packaging==23.2
pandas==2.1.4
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.0
phik==0.12.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
plotly==5.19.0
plotly-resampler==0.9.2
pmdarima==2.0.4
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.24.2
psutil==5.9.4
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycaret==3.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.0
pyee==11.0.0
Pygments==2.14.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyod==1.1.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.28.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.5.0
pywin32==305
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.10.3
replicate==0.13.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.5.2
rsa==4.9
safetensors==0.4.2
schemdraw==0.15
scikit-base==0.7.2
scikit-learn==1.4.1.post1
scikit-plot==0.3.7
scipy==1.11.3
seaborn==0.12.2
Send2Trash==1.8.0
sentence-transformers==2.5.1
six==1.16.0
sktime==0.26.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.1
streamlit==1.26.0
streamlit-option-menu==0.3.12
streamlit-pandas-profiling==0.1.3
streamlit-pdf-viewer==0.0.9
streamlit-webrtc==0.47.0
sympy==1.12
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
text2emotion==0.0.5
textblob==0.17.1
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
torch==2.2.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.38.2
tsdownsample==0.1.2
typeguard==4.1.5
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.16
validators==0.22.0
visions==0.7.5
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.3.7
whisper==1.1.10
widgetsnbextension==4.0.5
wordcloud==1.9.3
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.4
ydata-profiling==4.6.4
yellowbrick==1.5
youtube-comment-downloader==0.1.70
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
zipp==3.16.2
